+++ include/libtropic_common.h
@@ -893,8 +893,14 @@ typedef enum lt_ecc_slot_t {
     TR01_ECC_SLOT_31,
 } lt_ecc_slot_t;
 
-/** @brief ECC key type */
-typedef enum lt_ecc_curve_type_t { TR01_CURVE_P256 = 1, TR01_CURVE_ED25519 } lt_ecc_curve_type_t;
+/** @brief Type of the Elliptic Curve used for the key pair. Used by the ECC_Key_Generate,
+ * ECC_Key_Store and ECC_Key_Read L3 commands. */
+typedef enum lt_ecc_curve_type_t {
+    /** @brief Curve P-256. */
+    TR01_CURVE_P256 = 1,
+    /** @brief Curve Ed25519. */
+    TR01_CURVE_ED25519 = 2
+} lt_ecc_curve_type_t;
 
 /** @brief Length of public keys for P256 curve. */
 #define TR01_CURVE_P256_PUBKEY_LEN 64
@@ -903,8 +909,14 @@ typedef enum lt_ecc_curve_type_t { TR01_CURVE_P256 = 1, TR01_CURVE_ED25519 } lt_
 /** @brief Common length of private keys for both P256 and ED25519 curves. */
 #define TR01_CURVE_PRIVKEY_LEN 32
 
-/** @brief ECC key origin */
-typedef enum lt_ecc_key_origin_t { TR01_CURVE_GENERATED = 1, TR01_CURVE_STORED } lt_ecc_key_origin_t;
+/** @brief Origin of the Elliptic Curve private key used to derive the public key. Used by the
+ * ECC_Key_Read L3 command. */
+typedef enum lt_ecc_key_origin_t {
+    /** @brief Private key was generated via the ECC_Key_Generate L3 command. */
+    TR01_KEY_GENERATED = 1,
+    /** @brief Private key was stored via the ECC_Key_Store L3 command. */
+    TR01_KEY_STORED = 2
+} lt_ecc_key_origin_t;
 
 /** @brief Length of the EC signature (RS) for both ECDSA and EDDSA. */
 #define TR01_ECDSA_EDDSA_SIGNATURE_LENGTH 64